Refactor "query from request" builder
Minor refactoring on flask-restalchemy query builder. Also added more
tests for filters and order_by operator.

RFDAP-938
