+++ flask_restalchemy/tests/test_simple_api.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-import json
 
 import pytest
 
@@ -83,106 +82,3 @@ def test_put(client):
     emp3 = Employee.query.get(1)
     assert emp3.firstname == 'Jimmy'
 
-def test_filter(client):
-    for i in range(20):
-        client.post('/company', data={ 'name': 'Terrans {}'.format(i), 'location': 'Location {}'.format(20 - i)})
-
-    response = client.get('/company')
-    dataList = response.parsed_data
-    assert len(dataList) == 21
-
-    response = client.get('/company?limit=5')
-    dataList = response.parsed_data
-    assert len(dataList) == 5
-
-    response = client.get('/company?filter={}'.format(json.dumps({"name": "Terrans 1"})))
-    dataList = response.parsed_data
-    assert len(dataList) == 1
-
-    response = client.get('/company?filter={}'.format(json.dumps({"name": {"eq": "Terrans 1"}})))
-    dataList = response.parsed_data
-    assert len(dataList) == 1
-
-
-    response = client.get('/company?filter={}'.format(json.dumps({
-        "$or": {
-            "name": {"eq": "Terrans 1"},
-            "location": "Location 1"
-        }
-    })))
-    dataList = response.parsed_data
-    assert len(dataList) == 2
-
-    # test if AND is the default Logical Operator
-    response = client.get('/company?filter={}'.format(json.dumps({
-        "name": {"eq": "Terrans 1"},
-        "location": "Location 1"
-    })))
-    dataList = response.parsed_data
-    assert len(dataList) == 0
-
-    response = client.get('/company?filter={}'.format(json.dumps({
-        "$and": {
-            "name": {"eq": "Terrans 10"},
-            "location": "Location 10"
-        }
-    })))
-    dataList = response.parsed_data
-    assert len(dataList) == 1
-
-    response = client.get('/company?filter={}'.format(json.dumps({"name": { "in": ["Terrans 1", "Terrans 2"]}})))
-    dataList = response.parsed_data
-    assert len(dataList) == 2
-
-    response = client.get('/company?filter={}'.format(json.dumps({"name": {"endswith": '9'}})))
-    dataList = response.parsed_data
-    assert len(dataList) == 2
-
-    response = client.get('/company?filter={}'.format(json.dumps({"name": {"startswith": 'Terr'}})))
-    dataList = response.parsed_data
-    assert len(dataList) == 21
-
-    response = client.get('/company?limit=5&filter={}'.format(json.dumps({"name": {"startswith": 'Terr'}})))
-    dataList = response.parsed_data
-    assert len(dataList) == 5
-
-    with pytest.raises(ValueError, message='Unknown operator unknown_operator'):
-        client.get('/company?filter={}'.format(json.dumps({"name": {"unknown_operator": 'Terr'}})))
-
-
-def test_pagination(client):
-    for i in range(20):
-        client.post('/company', data={ 'name': 'Terrans {}'.format(i)})
-
-    response = client.get('/company?page=1&per_page=50')
-    dataList = response.parsed_data
-    assert len(dataList.get('results')) == 21
-
-    response = client.get('/company?page=1&per_page=5')
-    dataList = response.parsed_data
-    assert len(dataList.get('results')) == 5
-
-    response = client.get('/company?page=5&per_page=5')
-    dataList = response.parsed_data
-    assert len(dataList.get('results')) == 1
-
-
-def test_relations_pagination(client):
-    response = client.post('/company', data={'name': 'Terrans 1'})
-    assert response.status_code == 201
-    company_id = response.parsed_data['id']
-
-    for i in range(20):
-        client.post('/company/{}/employees'.format(company_id), data={'firstname': 'Jimmy {}'.format(i)})
-
-    response = client.get('/company/{}/employees?filter={}'.format(company_id, json.dumps({"firstname": {"eq": "Jimmy 1"}})))
-    assert response.status_code == 200
-    dataList = response.parsed_data
-    assert len(dataList) == 1
-    assert 'firstname' in dataList[0]
-    assert dataList[0]['firstname'] == 'Jimmy 1'
-
-    response = client.get('/company/{}/employees?page=1&per_page=5'.format(company_id))
-    assert response.status_code == 200
-    dataList = response.parsed_data
-    assert len(dataList.get('results')) == 5
